fix standalone bootstrap problems
