+++ modules/jb-module-audio/src/main/java/ru/caramel/juniperbot/module/audio/service/impl/PlaylistServiceImpl.java
@@ -16,13 +16,6 @@
  */
 package ru.caramel.juniperbot.module.audio.service.impl;
 
-import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler;
-import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
-import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayerManager;
-import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager;
-import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
-import com.sedmelluq.discord.lavaplayer.tools.io.MessageInput;
-import com.sedmelluq.discord.lavaplayer.tools.io.MessageOutput;
 import com.sedmelluq.discord.lavaplayer.track.*;
 import net.dv8tion.jda.core.entities.Member;
 import org.apache.commons.collections4.CollectionUtils;
@@ -30,39 +23,29 @@ import org.apache.commons.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.caramel.juniperbot.core.persistence.entity.LocalMember;
 import ru.caramel.juniperbot.core.service.*;
-import ru.caramel.juniperbot.module.audio.model.StoredPlaylist;
 import ru.caramel.juniperbot.module.audio.model.PlaybackInstance;
 import ru.caramel.juniperbot.module.audio.model.TrackRequest;
 import ru.caramel.juniperbot.module.audio.persistence.entity.Playlist;
 import ru.caramel.juniperbot.module.audio.persistence.entity.PlaylistItem;
 import ru.caramel.juniperbot.module.audio.persistence.repository.PlaylistItemRepository;
 import ru.caramel.juniperbot.module.audio.persistence.repository.PlaylistRepository;
-import ru.caramel.juniperbot.module.audio.service.LavaAudioService;
 import ru.caramel.juniperbot.module.audio.service.PlaylistService;
+import ru.caramel.juniperbot.module.audio.service.handling.JbAudioPlayerManager;
 import ru.caramel.juniperbot.module.audio.utils.PlaylistUtils;
 
-import javax.annotation.PostConstruct;
-import java.io.*;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import static com.sedmelluq.discord.lavaplayer.tools.FriendlyException.Severity.COMMON;
-
 @Service
-public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager {
+public class PlaylistServiceImpl implements PlaylistService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(PlaylistServiceImpl.class);
-
-    private final static String PLAYLIST_PATTERN = "https?:\\/\\/%s/playlist/([a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12})";
 
     @Autowired
     private PlaylistRepository playlistRepository;
@@ -71,162 +54,12 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
     private PlaylistItemRepository playlistItemRepository;
 
     @Autowired
-    private BrandingService brandingService;
-
-    @Autowired
-    private MessageService messageService;
-
-    @Autowired
-    private ContextService contextService;
-
-    @Autowired
-    private ConfigService configService;
-
-    @Autowired
     private MemberService memberService;
 
     @Autowired
-    private LavaAudioService lavaAudioService;
+    private ApplicationContext applicationContext;
 
-    private Method sourceMethod;
-
-    @PostConstruct
-    public void init() {
-        try {
-            sourceMethod = DefaultAudioPlayerManager.class.getDeclaredMethod("checkSourcesForItem",
-                    AudioReference.class, AudioLoadResultHandler.class, boolean[].class);
-            sourceMethod.setAccessible(true);
-        } catch (NoSuchMethodException e) {
-            LOGGER.warn("Could not get LavaPlayer init method");
-        }
-    }
-
-    @Override
-    @Transactional
-    public AudioItem loadItem(DefaultAudioPlayerManager manager, AudioReference reference) {
-        Pattern pattern = Pattern.compile(String.format(PLAYLIST_PATTERN, Pattern.quote(brandingService.getWebHost())));
-        Matcher matcher = pattern.matcher(reference.identifier);
-        if (matcher.find()) {
-            String uuid = matcher.group(1);
-            Playlist playlist = getPlaylist(uuid);
-            if (playlist != null) {
-                List<PlaylistItem> items = new ArrayList<>(playlist.getItems());
-                if (CollectionUtils.isNotEmpty(items)) {
-                    List<AudioTrack> tracks = new ArrayList<>(items.size());
-                    items.forEach(e -> {
-                        try {
-                            AudioTrack track = getTracksFor(manager, e);
-                            if (track != null) {
-                                tracks.add(track);
-                            }
-                        } catch (Exception ex) {
-                            // fall down and ignore it
-                        }
-                    });
-                    if (!tracks.isEmpty()) {
-                        refreshStoredPlaylist(playlist, tracks);
-                        return new StoredPlaylist(playlist, tracks);
-                    }
-                }
-            }
-            onError(playlist, "discord.command.audio.playlist.notFound");
-        }
-        return null;
-    }
-
-    private void onError(Playlist playlist, String messageCode) throws FriendlyException {
-        contextService.withContext(playlist.getGuildId(), () -> {
-            throw new FriendlyException(messageService.getMessage(messageCode), COMMON, null);
-        });
-    }
-
-    private AudioTrack getTracksFor(DefaultAudioPlayerManager manager, PlaylistItem item) {
-        if (sourceMethod == null) {
-            return null;
-        }
-
-        class TrackHolder {
-            AudioTrack track;
-        }
-
-        TrackHolder holder = new TrackHolder();
-
-        if (ArrayUtils.isNotEmpty(item.getData())) {
-            holder.track = decodeTrack(item.getData());
-        }
-
-        if (holder.track == null) {
-            AudioLoadResultHandler handler = new AudioLoadResultHandler() {
-                @Override
-                public void trackLoaded(AudioTrack track) {
-                    holder.track = track;
-                }
-
-                @Override
-                public void playlistLoaded(AudioPlaylist playlist) {
-                    // no playlist supported
-                }
-
-                @Override
-                public void noMatches() {
-                    // nothing
-                }
-
-                @Override
-                public void loadFailed(FriendlyException exception) {
-                    // nothing
-                }
-            };
-
-            if (!checkSourcesForItem(manager, new AudioReference(item.getUri(), item.getTitle()), handler)) {
-                checkSourcesForItem(manager, new AudioReference(item.getIdentifier(), item.getTitle()), handler);
-            }
-
-            if (ArrayUtils.isEmpty(item.getData()) && holder.track != null) {
-                item.setData(encodeTrack(holder.track));
-                playlistItemRepository.save(item);
-            }
-        }
-        return holder.track;
-    }
-
-    private boolean checkSourcesForItem(DefaultAudioPlayerManager manager, AudioReference reference, AudioLoadResultHandler resultHandler) {
-        if (sourceMethod == null) {
-            return false;
-        }
-        try {
-            Object result = sourceMethod.invoke(manager, reference, resultHandler, new boolean[1]);
-            return Boolean.TRUE.equals(result);
-        } catch (IllegalAccessException | InvocationTargetException e) {
-            LOGGER.warn("Could not invoke checkSourcesForItem of LavaPlayer");
-            return false;
-        }
-    }
-
-    @Override
-    public AudioTrack decodeTrack(AudioTrackInfo trackInfo, DataInput input) throws IOException {
-        return null;
-    }
-
-    @Override
-    public boolean isTrackEncodable(AudioTrack track) {
-        return true;
-    }
-
-    @Override
-    public void encodeTrack(AudioTrack track, DataOutput output) throws IOException {
-        // No custom values that need saving
-    }
-
-    @Override
-    public void shutdown() {
-        // nothing to shutdown
-    }
-
-    @Override
-    public String getSourceName() {
-        return "jbPlaylist";
-    }
+    private JbAudioPlayerManager audioPlayerManager;
 
     @Override
     @Transactional
@@ -238,6 +71,18 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
     @Transactional
     public Playlist find(Long id) {
         return validateItems(playlistRepository.findById(id).orElse(null));
+    }
+
+    @Override
+    @Transactional
+    public Playlist save(Playlist playlist) {
+        return playlistRepository.save(playlist);
+    }
+
+    @Override
+    @Transactional
+    public PlaylistItem save(PlaylistItem playlistItem) {
+        return playlistItemRepository.save(playlistItem);
     }
 
     @Override
@@ -282,7 +127,7 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
                 for (TrackRequest request : requests) {
                     PlaylistItem item = new PlaylistItem(request.getTrack(), localMember);
                     item.setPlaylist(playlist);
-                    item.setData(encodeTrack(request.getTrack()));
+                    item.setData(getAudioPlayerManager().encodeTrack(request.getTrack()));
                     playlist.getItems().add(item);
                 }
                 playlistRepository.save(playlist);
@@ -292,9 +137,9 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
         }
     }
 
-    @Override
-    @Transactional
     @Async
+    @Transactional
+    @Override
     public void refreshStoredPlaylist(PlaybackInstance instance) {
         try {
             Playlist playlist = getPlaylist(instance);
@@ -308,7 +153,7 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
                     item.setPlaylist(playlist);
                 }
                 if (ArrayUtils.isEmpty(item.getData())) {
-                    item.setData(encodeTrack(e.getTrack()));
+                    item.setData(getAudioPlayerManager().encodeTrack(e.getTrack()));
                 }
                 newItems.add(item);
             });
@@ -323,7 +168,9 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
         }
     }
 
-    private void refreshStoredPlaylist(Playlist playlist, List<AudioTrack> tracks) {
+    @Transactional
+    @Override
+    public void refreshStoredPlaylist(Playlist playlist, List<AudioTrack> tracks) {
         try {
             List<PlaylistItem> toRemove = new ArrayList<>(playlist.getItems());
             List<PlaylistItem> existentItems = tracks.stream()
@@ -341,40 +188,18 @@ public class PlaylistServiceImpl implements PlaylistService, AudioSourceManager 
         }
     }
 
-    private byte[] encodeTrack(AudioTrack track) {
-        if (track == null) {
-            return null;
-        }
-        AudioPlayerManager manager = lavaAudioService.getPlayerManager();
-        try (ByteArrayOutputStream outputStream = new ByteArrayOutputStream()) {
-            MessageOutput output = new MessageOutput(outputStream);
-            manager.encodeTrack(output, track);
-            return outputStream.toByteArray();
-        } catch (IOException e) {
-            LOGGER.warn("Could not encode track {}", track);
-        }
-        return null;
-    }
-
-    private AudioTrack decodeTrack(byte[] data) {
-        if (ArrayUtils.isEmpty(data)) {
-            return null;
-        }
-        AudioPlayerManager manager = lavaAudioService.getPlayerManager();
-        try (ByteArrayInputStream stream = new ByteArrayInputStream(data)) {
-            DecodedTrackHolder holder = manager.decodeTrack(new MessageInput(stream));
-            return holder != null && holder.decodedTrack != null ? holder.decodedTrack : null;
-        } catch (IOException e) {
-            LOGGER.warn("Could not decode track");
-        }
-        return null;
-    }
-
     private Playlist validateItems(Playlist playlist) {
         if (playlist != null && CollectionUtils.isNotEmpty(playlist.getItems()) && playlist.getItems().contains(null)) {
             playlist.setItems(playlist.getItems().stream().filter(Objects::nonNull).collect(Collectors.toList()));
             playlistRepository.save(playlist);
         }
         return playlist;
+    }
+
+    private JbAudioPlayerManager getAudioPlayerManager() {
+        if (audioPlayerManager == null) {
+            audioPlayerManager = applicationContext.getBean(JbAudioPlayerManager.class);
+        }
+        return audioPlayerManager;
     }
 }
